Problem.
Could't get password generator to function. Going a different route. Found youtube video on how to build generator that still fits the criteria. going to try and us it as an outline as deadline approches. Good news though, is that it uses DOM elemets which are relevant to the most recent week.