Added a css and html
Added CSS stylesheet and reset to repository. Same with html. Followed the tutorial to make sure the lablinging was consistent, but it doesnt seem to work. Out of time. Will be personalising  in the next few days. Wish that I could have done better on this one. Turning in something is better than nothing though I guess.